services/scheduler-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3
httpx==0.25.0
pytest==7.4.3
pytest-asyncio==0.21.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
loguru==0.7.2
aiocache==0.12.2
ortools==9.8.3296
numpy==1.26.1
pandas==2.1.1
celery==5.3.4
redis==5.0.1
tenacity==8.2.3
deap==1.4.1  # For genetic algorithms
